REF: Consolidate Results and Utilities under experiments module
