+++ skordinal/experiments/tests/__init__.py
@@ -1,0 +1,3 @@
+"""Tests for the experiments module."""
+
+__all__: list[str] = []
