--- skordinal/results/__init__.py
@@ -1,7 +1,0 @@
-"""Results handling module."""
-
-from ._results import Results
-
-__all__ = [
-    "Results",
-]
