--- skordinal/results/tests/__init__.py
@@ -1,3 +1,0 @@
-"""Tests for results handling module."""
-
-__all__: list[str] = []
