--- skordinal/utilities/__init__.py
@@ -1,5 +1,0 @@
-"""Utilities module."""
-
-from ._utilities import Utilities
-
-__all__ = ["Utilities"]
