--- skordinal/utilities/tests/__init__.py
@@ -1,3 +1,0 @@
-"""Tests for experiment handling."""
-
-__all__: list[str] = []
